+++ src/templates/clazzDTO.ts
@@ -1,9 +1,0 @@
-import CreateComponent from './interfaces/CreateComponent';
-import pascalCase from './shared/functions/pascal-case';
-
-export default ({ componentName }: CreateComponent) => (
-`class ${ pascalCase(componentName) }DTO {
-
-}
-`
-);