+++ src/templates/clazzException.ts
@@ -1,0 +1,11 @@
+import CreateComponent from './interfaces/CreateComponent';
+import pascalCase from './shared/functions/pascal-case';
+
+export default ({ componentName }: CreateComponent) => (
+`class ${ pascalCase(componentName) }Exception implements Exception {
+  final String? message;
+
+  ${ pascalCase(componentName) }Exception([this.message]);
+}
+`
+);