+++ src/templates/statelessWidget.ts
@@ -9,7 +9,12 @@ class ${ pascalCase(componentName) } extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Container();
+    return const Center(
+      child: Text(
+        '${ pascalCase(componentName) } is working',
+        style: TextStyle(fontSize: 20),
+      ),
+    );
   }
 }
 `